requirements.txt
selenium==3.141.0
mysql-connector-python==8.1.0
newspaper3k==0.2.8
beautifulsoup4==4.12.2
beautifulsoup4==4.12.2
